Added the cart view (#97)
* Added the Cart view to the project as well as the ability to remove items from the cart.

It doesn't look identical to the figma but its pretty close.

* added cart to Store

This works mostly. There is one issue with the app freezing after the reducer functions are called. This only happens if the functions are called after the useSelector function though.

* fix recursion bug

* Added the cart merchant to the stored cart

The merchant the cart items are from is now stored in cart. Also changed the default screen back to home.

* making it prettier

ran 'npx prettier --write .'

Co-authored-by: Jordan Ribbink <jribbink@telus.net>