Added a working Map
Restaurants with a location object will appear on the map now and are interactable. Small issue with the action sheet showing up in front of the navigation bar (https://github.com/gorhom/react-native-portal) may be useful. The default start location for the view is currently stored in some variable called userLocation. There is some small things that are still needed but the overall map is mostly done.