Added geocoding (kinda)
The user location is now retrieved and is stored in userLocation variable but not before the map centers on the default location. So its a small issue that needs to be fixed. I tried making it into a hook but the async/await functions were giving me some issues. But the necessary code is there for finding the location.