Riley comer/map (#126)
* Added a working Map

Restaurants with a location object will appear on the map now and are interactable. Small issue with the action sheet showing up in front of the navigation bar (https://github.com/gorhom/react-native-portal) may be useful. The default start location for the view is currently stored in some variable called userLocation. There is some small things that are still needed but the overall map is mostly done.

* Added geocoding (kinda)

The user location is now retrieved and is stored in userLocation variable but not before the map centers on the default location. So its a small issue that needs to be fixed. I tried making it into a hook but the async/await functions were giving me some issues. But the necessary code is there for finding the location.

* ran 'npx prettier --write .'

Co-authored-by: Jordan Colledge <71358703+ItsMyFuneral@users.noreply.github.com>