Endpoints and backend done.
Backend should work. (might need to be adjusted for real time updates). Front end has a basic inputs on account page for changing merchant details. App needs to be refreshed to see changes on all pages at moment. The main problem at the moment is image inputs. I tried using image picker library but I was not able to get it to work. I have an image update endpoint that should work but have not been able to test it due to lack of inputs.